refactor(tibuild): move `pkg/database` to `internal/database`
