+++ tibuild/api/api.go
@@ -11,8 +11,8 @@ import (
 	ginSwagger "github.com/swaggo/gin-swagger"
 
 	"github.com/PingCAP-QE/ee-apps/tibuild/internal/controller"
+	"github.com/PingCAP-QE/ee-apps/tibuild/internal/database"
 	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/configs"
-	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/database"
 	controllers "github.com/PingCAP-QE/ee-apps/tibuild/pkg/rest/controller"
 	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/rest/service"
 	events "github.com/PingCAP-QE/ee-apps/tibuild/pkg/webhook/handler"
