+++ tibuild/internal/controller/pipeline_trigger.go
@@ -11,10 +11,10 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/rs/zerolog/log"
 
+	"github.com/PingCAP-QE/ee-apps/tibuild/internal/database"
 	"github.com/PingCAP-QE/ee-apps/tibuild/internal/entity"
 	"github.com/PingCAP-QE/ee-apps/tibuild/internal/service"
 	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/configs"
-	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/database"
 )
 
 const (
