+++ tibuild/internal/controller/request_rotation.go
@@ -6,8 +6,8 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
 
+	"github.com/PingCAP-QE/ee-apps/tibuild/internal/database"
 	"github.com/PingCAP-QE/ee-apps/tibuild/internal/entity"
-	"github.com/PingCAP-QE/ee-apps/tibuild/pkg/database"
 )
 
 type RequestRotationRequestStruct struct {
